+++ client.go
@@ -1,6 +1,7 @@
 package cache
 
 import (
+	"context"
 	"reflect"
 	"time"
 )
@@ -10,29 +11,29 @@ type Client interface {
 	// Get gets the item for the given key.
 	// Returns nil for a cache miss.
 	// The key must be at most 250 bytes in length.
-	Get(key string, data interface{}) error
+	Get(ctx context.Context, key string, data interface{}) error
 
 	// Set writes the given item, unconditionally.
-	Set(key string, data interface{}, expiration time.Time) error
+	Set(ctx context.Context, key string, data interface{}, expiration time.Time) error
 
 	// Add writes the given item, if no value already exists for its
 	// key. ErrNotStored is returned if that condition is not met.
-	Add(key string, data interface{}, expiration time.Time) error
+	Add(ctx context.Context, key string, data interface{}, expiration time.Time) error
 
 	// Delete deletes the item with the provided key, if it exists.
-	Delete(key string) error
+	Delete(ctx context.Context, key string) error
 
 	// Increment adds delta to the currently stored number
 	// If the key does not exist, it will be initialized to 0 with no expiration.
 	// If the value overflows, it will loop around from 0
 	// For many client implementations, you need to be using a LiteralEncoding for this feature to work
-	Increment(key string, delta uint64) (uint64, error)
+	Increment(ctx context.Context, key string, delta uint64) (uint64, error)
 
 	// Decrement subtracts delta to the currently stored number
 	// If the key does not exist, it will be initialized to 0 with no expiration, which will make it overflow.
 	// If the value overflows, it will loop around from MaxUint64
 	// For many client implementations, you need to be using a LiteralEncoding for this feature to work
-	Decrement(key string, delta uint64) (uint64, error)
+	Decrement(ctx context.Context, key string, delta uint64) (uint64, error)
 }
 
 var NeverExpire time.Time