+++ client_test.go
@@ -1,6 +1,7 @@
 package cache
 
 import (
+	"context"
 	"fmt"
 	"math"
 	"math/rand"
@@ -12,11 +13,12 @@ import (
 
 func testClient(t *testing.T, client Client, encoding Encoding) {
 	r := rand.NewSource(time.Now().UnixNano())
+	ctx := context.Background()
 
 	t.Run("get not existing", func(t *testing.T) {
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 		var data interface{}
-		err := client.Get(testKey, &data)
+		err := client.Get(ctx, testKey, &data)
 		require.Zero(t, data)
 		require.EqualError(t, err, "cache miss")
 	})
@@ -24,11 +26,11 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 	t.Run("set", func(t *testing.T) {
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 		for _, data := range []int{123, 124} {
-			err := client.Set(testKey, data, time.Now().Add(1*time.Second))
+			err := client.Set(ctx, testKey, data, time.Now().Add(1*time.Second))
 			require.NoError(t, err)
 
 			var loaded int
-			err = client.Get(testKey, &loaded)
+			err = client.Get(ctx, testKey, &loaded)
 			require.NoError(t, err)
 			require.Equal(t, data, loaded)
 		}
@@ -38,20 +40,20 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 		data := 123
 
-		err := client.Add(testKey, data, time.Now().Add(1*time.Second))
+		err := client.Add(ctx, testKey, data, time.Now().Add(1*time.Second))
 		require.NoError(t, err)
 
 		var loaded int
-		err = client.Get(testKey, &loaded)
+		err = client.Get(ctx, testKey, &loaded)
 		require.NoError(t, err)
 		require.Equal(t, data, loaded)
 
 		data2 := 124
-		err = client.Add(testKey, data2, time.Now().Add(1*time.Second))
+		err = client.Add(ctx, testKey, data2, time.Now().Add(1*time.Second))
 		require.EqualError(t, err, "not stored")
 
 		var loaded2 int
-		err = client.Get(testKey, &loaded2)
+		err = client.Get(ctx, testKey, &loaded2)
 		require.NoError(t, err)
 		require.Equal(t, data, loaded2)
 	})
@@ -60,19 +62,19 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 		stored := 123
 
-		err := client.Set(testKey, stored, time.Now().Add(1*time.Second))
+		err := client.Set(ctx, testKey, stored, time.Now().Add(1*time.Second))
 		require.NoError(t, err)
 
 		var loaded int
-		err = client.Get(testKey, &loaded)
+		err = client.Get(ctx, testKey, &loaded)
 		require.NoError(t, err)
 		require.Equal(t, loaded, loaded)
 
-		err = client.Delete(testKey)
+		err = client.Delete(ctx, testKey)
 		require.NoError(t, err)
 
 		var loaded2 int
-		err = client.Get(testKey, &loaded2)
+		err = client.Get(ctx, testKey, &loaded2)
 		require.Zero(t, loaded2)
 		require.EqualError(t, err, "cache miss")
 	})
@@ -85,18 +87,18 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 		stored := 123
 
-		err := client.Set(testKey, stored, time.Now().Add(1*time.Second))
+		err := client.Set(ctx, testKey, stored, time.Now().Add(1*time.Second))
 		require.NoError(t, err)
 
 		var loaded int
-		err = client.Get(testKey, &loaded)
+		err = client.Get(ctx, testKey, &loaded)
 		require.NoError(t, err)
 		require.Equal(t, loaded, loaded)
 
 		time.Sleep(1500 * time.Millisecond)
 
 		var loaded2 int
-		err = client.Get(testKey, &loaded2)
+		err = client.Get(ctx, testKey, &loaded2)
 		require.Zero(t, loaded2)
 		require.EqualError(t, err, "cache miss")
 	})
@@ -107,11 +109,11 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 		}
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 
-		newVal, err := client.Increment(testKey, 123)
+		newVal, err := client.Increment(ctx, testKey, 123)
 		require.NoError(t, err)
 		require.Equal(t, uint64(123), newVal)
 
-		newVal, err = client.Increment(testKey, 10)
+		newVal, err = client.Increment(ctx, testKey, 10)
 		require.NoError(t, err)
 		require.Equal(t, uint64(133), newVal)
 	})
@@ -122,11 +124,11 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 		}
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 
-		newVal, err := client.Increment(testKey, uint64(math.MaxUint64))
+		newVal, err := client.Increment(ctx, testKey, uint64(math.MaxUint64))
 		require.NoError(t, err)
 		require.Equal(t, uint64(math.MaxUint64), newVal)
 
-		newVal, err = client.Increment(testKey, 10)
+		newVal, err = client.Increment(ctx, testKey, 10)
 		require.NoError(t, err)
 		require.Equal(t, uint64(9), newVal)
 	})
@@ -137,10 +139,10 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 		}
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 
-		err := client.Set(testKey, uint64(123), time.Now().Add(1*time.Second))
+		err := client.Set(ctx, testKey, uint64(123), time.Now().Add(1*time.Second))
 		require.NoError(t, err)
 
-		newVal, err := client.Decrement(testKey, 10)
+		newVal, err := client.Decrement(ctx, testKey, 10)
 		require.NoError(t, err)
 		require.Equal(t, uint64(113), newVal)
 	})
@@ -151,7 +153,7 @@ func testClient(t *testing.T, client Client, encoding Encoding) {
 		}
 		testKey := fmt.Sprintf("go-cache-test-%d", r.Int63())
 
-		newVal, err := client.Decrement(testKey, 10)
+		newVal, err := client.Decrement(ctx, testKey, 10)
 		require.NoError(t, err)
 		require.Equal(t, math.MaxUint64-uint64(9), newVal)
 	})